Require only the minimum number of user properties to be specified.
This is done in two ways: for properties that must be
provided (active, allowed_ip_list, authentication_provider_id,
locale_id and roles), we allow a default value to be configured at the
team level. If a user does not have a value explicitly specified for
any of these properties, the team default is used.

For optional properties, if not provided, a null value is used.