Move user information to a separate file.
One complaint with the previous implementation was that having the
user information in each team was repetitive and error prone. To
address this, the user information is now written to (and read from) a
separate file, users.yml, while the teams files now contain just
references to their user members (a reference consisting of the
authentication provider name and the username).

The logging configuration can now be specified in a configuration
file.

Unit tests have been added.