Handle duplicate teams and users via the validate method.
Previously, we detected duplicate teams and users in the Model
constructor. It is more consistent if we detect these cases in the
validate() method. However, when retrieving model data from Access
Control, it is implicitly valid, so an optional constructor argument
has been added to allow the relevant processing to occur in the
constructor.